requirements.txt
django==5.1.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.4.0
channels==4.1.0
dj-database-url==2.1.0
python-dotenv==1.0.1
Pillow==10.4.0
psycopg2-binary==2.9.10

